+++ src/diagnostics/GtpCommandViewController.m
@@ -108,7 +108,10 @@
 {
   UITableViewCell* cell = [TableViewCellFactory cellWithType:DefaultCellType tableView:tableView];
   cell.accessoryType = UITableViewCellAccessoryDisclosureIndicator;
-  cell.textLabel.text = [self.model commandStringAtIndex:indexPath.row];
+  // Cast is required because NSInteger and int differ in size in 64-bit. Cast
+  // is safe because this app was not made to handle more than pow(2, 31)
+  // commands.
+  cell.textLabel.text = [self.model commandStringAtIndex:(int)indexPath.row];
   cell.showsReorderControl = YES;
   return cell;
 }
@@ -122,7 +125,10 @@
   {
     case UITableViewCellEditingStyleDelete:
     {
-      [self.model removeCommandAtIndex:indexPath.row];
+      // Cast is required because NSInteger and int differ in size in 64-bit.
+      // Cast is safe because this app was not made to handle more than
+      // pow(2, 31) commands.
+      [self.model removeCommandAtIndex:(int)indexPath.row];
       [tableView deleteRowsAtIndexPaths:[NSArray arrayWithObject:indexPath]
                        withRowAnimation:UITableViewRowAnimationRight];
       break;
@@ -137,7 +143,10 @@
 // -----------------------------------------------------------------------------
 - (void) tableView:(UITableView*)tableView moveRowAtIndexPath:(NSIndexPath*)fromIndexPath toIndexPath:(NSIndexPath*)toIndexPath
 {
-  [self.model moveCommandAtIndex:fromIndexPath.row toIndex:toIndexPath.row];
+  // Cast is required because NSInteger and int differ in size in 64-bit. Cast
+  // is safe because this app was not made to handle more than pow(2, 31)
+  // commands.
+  [self.model moveCommandAtIndex:(int)fromIndexPath.row toIndex:(int)toIndexPath.row];
 }
 
 #pragma mark - UITableViewDelegate overrides
@@ -148,7 +157,10 @@
 - (void) tableView:(UITableView*)tableView didSelectRowAtIndexPath:(NSIndexPath*)indexPath
 {
   [tableView deselectRowAtIndexPath:indexPath animated:NO];
-  [self editCommandAtIndex:indexPath.row];
+  // Cast is required because NSInteger and int differ in size in 64-bit. Cast
+  // is safe because this app was not made to handle more than pow(2, 31)
+  // commands.
+  [self editCommandAtIndex:(int)indexPath.row];
 }
 
 #pragma mark - Action handlers
